combined case info and hire date
